Changed default op_life for trn to 60
The default operational life for transmission technologies (hardcoded) was set to 30 years whereas the default operational life for transmission technologies introduced through the user configuration file was set to 30 leading to discrepancies. I've set the default to 60 for both cases since this is more in line with real world application.