Cost and OL updates
Since in earlier versions capital costs was only assigned to CapitalCostStorage it meant that the model could freely build (not use) PWR capacity up to the max cap limit (which it did for all techs and years). I.e. this led to a situation where the full storage capacity could be charged/discharged within a single timeslice. Changes are made to split capital costs to CapitalCost (50%) and CapitalCostStorage (50%). OperationalLife has also been added alongside OperationalLifeStorage.