Update to LINEAR emission limit calculations
Previously historical emission data from EMBER was only pulled if the historical year was within the model horizon. However, this meant that in certain instances no LINEAR interpolation could be applied. Changes have been made to pull the latest historical data year, perform the interpolation, and drop historical years that are out of the model horizon.